refactor(CarGameController): 자동차 경주 게임을 진행하는 클래스 분리
- play라는 메서드를 통해 게임 진행자가 게임을 수행하는 역할을 명시
